feat: Changed architecture
So instead of genius math I decided to go with foolproof approach. I will write the clauses into a tmp file and the vars are stored in a dictionary.

This allows me to generate everything, enumerate it and then write it into the final output file. That way I do not have to be correct at math.